% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{canopysnowint}
\alias{canopysnowint}
\title{Calculates canopy snow interception}
\usage{
canopysnowint(
  precd,
  uz,
  dtm,
  gsnowd,
  Lt,
  pai,
  plai,
  hgt,
  d = NA,
  zm = NA,
  phs = 375,
  Sh = 6.3,
  zu = 2
)
}
\arguments{
\item{uz}{wind speed at height \code{zu}  (m/s)}

\item{dtm}{a raster of elevations}

\item{gsnowd}{ground snow depth (cm)}

\item{Lt}{snow load in previous time step (mm SWE)}

\item{pai}{plant area index}

\item{plai}{proportion of \code{pai} that is living vegetation (i.e. leaves)}

\item{hgt}{canopy height (m)}

\item{d}{optionally, zero plane displacement height of wind profile}

\item{zm}{optionally, roughness length for momentum transfer}

\item{phs}{optionally, fresh snow density (kg / m^3)}

\item{Sh}{optionally, branch snow load coefficient (kg/m^2). 6.6 for pine and 5.9 kg for spruce.}

\item{zu}{height above ground of wind speed measurement \code{uz} (m)}

\item{prec}{daily precipitation (mm)}
}
\value{
\code{Int} snow water equivalent (mm) intercepted by canopy
}
\description{
Function \code{canopysnowint} calculates snow intercepted by the canopy
in a single time step using the model of Hedstrom and Pomeroy (1998) Hydrological
Processes, 12: 1611-1625.
}
