% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climdata}
\alias{climdata}
\title{A data frame of hourly weather}
\format{
a data frame with the following elements:
\describe{
\item{obs_time}{POSIXlt object of dates and times in UTC}
\item{temp}{temperature (degrees C)}
\item{relhum}{relative humidity (percentage)}
\item{pres}{atmospheric press (kPa)}
\item{swrad}{Total incoming shortwave radiation (W / m^2)}
\item{difrad}{Diffuse radiation (W / m^2)}
\item{skyem}{Sky emissivity (0-1)}
\item{windspeed}{Wind speed (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
}
}
\source{
\url{https://en.ilmatieteenlaitos.fi/}
}
\usage{
climdata
}
\description{
A data frame of hourly weather (2nd Oct 2017 to 1st Apr 2019) for Sodankylä in Finland (67.367N, 26.629E)
}
\keyword{datasets}
