% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.freeze}
\alias{.freeze}
\title{Sets snow temperature to zero when it freezes to account for latent heat of freezing}
\usage{
.freeze(snowtemp)
}
\description{
Sets snow temperature to zero when it freezes to account for latent heat of freezing
}
