% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.gcanopy}
\alias{.gcanopy}
\title{Calculates turbulent molar conductivity within canopy}
\usage{
.gcanopy(uf, h, d, z1, z0 = 0.0122, phi_h = 1, gmin = 0.05)
}
\description{
Calculates turbulent molar conductivity within canopy
}
