% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.gHaC}
\alias{.gHaC}
\title{Calculates thermal conductivity to canopy snow surface}
\usage{
.gHaC(u2, zu, snowdepth, dtm, hgta, pai, zmin, weather, dint, xyf)
}
\description{
Calculates thermal conductivity to canopy snow surface
}
