% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.gHaG}
\alias{.gHaG}
\title{Calculates thermal conductivity to ground snow surface}
\usage{
.gHaG(u2, zu, snowdepth, dtm, hgta, pai, zmin, weather)
}
\description{
Calculates thermal conductivity to ground snow surface
}
