% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.horizon2}
\alias{.horizon2}
\title{Calculate horizon angle above a given height (for wind sheltering) for all time steps}
\usage{
.horizon2(dtm, azimuth, hgt)
}
\description{
Calculate horizon angle above a given height (for wind sheltering) for all time steps
}
