% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.roughresample}
\alias{.roughresample}
\title{resamples roughness lengths (or zero-plane displacement) by xyf to smooth}
\usage{
.roughresample(zm, dtm, xyf, tint = NA)
}
\description{
resamples roughness lengths (or zero-plane displacement) by xyf to smooth
}
