% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.snowmelt}
\alias{.snowmelt}
\title{Computes snow melt when snow is an array}
\usage{
.snowmelt(
  weather,
  precd,
  dtm,
  lat,
  long,
  pai,
  x,
  hgt,
  hgta,
  snowdepth,
  snowenv = "Taiga",
  meltfact = 0.115,
  STparams,
  snowem = 0.99,
  zu = 2,
  zmin = 0.002,
  umin = 0.5,
  astc = 1.5,
  xyf = 10,
  initdepth = 0,
  dint = 24,
  slr = NA,
  apr = NA,
  clump
)
}
\description{
Computes snow melt when snow is an array
}
