% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{.snowPenMont}
\alias{.snowPenMont}
\title{Snow Penman-Monteith}
\usage{
.snowPenMont(
  tc,
  pk,
  ea,
  radabs,
  gHa,
  G,
  es = NA,
  tdew = NA,
  T_est = NA,
  allout = TRUE
)
}
\description{
Snow Penman-Monteith
}
