% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.snowrad}
\alias{.snowrad}
\title{Calculates radiation absorbed by snow surface}
\usage{
.snowrad(
  weather,
  dtm,
  hor,
  lat,
  long,
  epai,
  x,
  snowalb,
  snowem,
  merid = 0,
  dst = 0
)
}
\description{
Calculates radiation absorbed by snow surface
}
