% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{.snowsoiltemp_hr}
\alias{.snowsoiltemp_hr}
\title{Snow hourly surface tmep function}
\usage{
.snowsoiltemp_hr(
  microsnow,
  reqhgt = 0.05,
  pai_a = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA
)
}
\description{
Snow hourly surface tmep function
}
