% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{.snowtwostream}
\alias{.snowtwostream}
\title{Snow two-stream radiation function}
\usage{
.snowtwostream(
  microsnow,
  reqhgt = 0.05,
  pai_a = NA,
  slr = NA,
  apr = NA,
  hor = NA
)
}
\description{
Snow two-stream radiation function
}
