% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{.snowwind}
\alias{.snowwind}
\title{Snow wind function}
\usage{
.snowwind(
  microsnow,
  reqhgt = 0.05,
  pai_a = NA,
  xyf = 1,
  zf = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  gmn = 0.3
)
}
\description{
Snow wind function
}
