% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfworker.R
\name{.SourceD}
\alias{.SourceD}
\title{Calculates source concentration}
\usage{
.SourceD(
  reqhgt,
  swabs,
  tcan,
  tground,
  tref,
  skyem,
  pai,
  pai_a,
  clump,
  Hratio,
  folden
)
}
\description{
Calculates source concentration
}
