% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.twostreamparams}
\alias{.twostreamparams}
\title{Calculate parameters of two-stream radiation model}
\usage{
.twostreamparams(pait, x, clump, lref, gref, ltra, alt, si)
}
\description{
Calculate parameters of two-stream radiation model
}
