% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.unpackpai}
\alias{.unpackpai}
\title{Converts pai to an array if provided a single value or raster}
\usage{
.unpackpai(pai, n, dtm)
}
\description{
Converts pai to an array if provided a single value or raster
}
