% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.winddaily}
\alias{.winddaily}
\title{Calculates average daily wind speed so that canopy snow interception can be computed}
\usage{
.winddaily(weather)
}
\description{
Calculates average daily wind speed so that canopy snow interception can be computed
}
