% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtm}
\alias{dtm}
\title{A 25 m resolution digital terrain dataset.}
\format{
A PackedSpatRaster object with 50 rows and 50 columns
}
\source{
\url{https://aws.amazon.com/public-datasets/terrain/}
}
\usage{
dtm
}
\description{
A dataset of elevations (m) for Sodankylä in Finland (xmin = 2963700, xmax = 2964950,
ymin = 10221300, ymax = 10222550, CRS: 3395).
}
\keyword{datasets}
