% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowfitfunctions.R
\name{fitsnowtemp}
\alias{fitsnowtemp}
\title{The function \code{fitsnowtemp} derives model parameters for fitting the snow temperature model}
\usage{
fitsnowtemp(
  weather,
  precd,
  snowdepth,
  snowtemp,
  snowem = 0.99,
  zm = 0.002,
  umin = 0.5,
  astc = 1.5,
  plotout = FALSE
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{precd}{a vector of daily precipitation (mm).}

\item{snowdepth}{a vector of snow depth measurements (cm) corresponding
to each time period for which snow temperature measurements are available (can be interpolated)}

\item{snowtemp}{an vector of snow temperature (deg C) measurements (see details)}

\item{snowem}{optionally, numeric value of snow emissivity}

\item{zm}{optionally, numeric value of roughness length for momentum transfer of snow surface (m)}

\item{umin}{optionally, numeric value indicating minimum wind speed used in conductivity calculations (m/s)}

\item{astc}{optionally, numeric value indicating the temperature at which precipitation falls as snow (deg C)}

\item{plotout}{optional logical indicating whether to plot fitted relationship}
}
\value{
an object of class \code{SnowTparams} used for fitting the snow temperature model
}
\description{
The function \code{fitsnowtemp} derives model parameters for fitting the snow temperature model
}
\details{
The format and and units of \code{weather} must follow that in the example
dataset \code{climdata}. However, entries do not need to be hourly. They canshould correspond to
the time for which snow temperature measurements are available.
}
\examples{
require(NicheMapR)
# Derive estimates of snow temperature (and depth) using NicheMapR
nmrout<-runNMRSnow(weather, precd, 67.367, 26.629, snowenv = "Taiga", ALTT = 199)
# Use NicheMapR outputs as inputs to fitsnowtemp
STparams<-fitsnowtemp(climdata,precd,nmrout$SNOWDEP,nmrout$SNOWTEMP,plotout=TRUE)
}
