% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowfitfunctions.R
\name{getmeltf}
\alias{getmeltf}
\title{Returns inbuilt snow melt factors for different snow environments}
\usage{
getmeltf(snowenv = "Taiga")
}
\arguments{
\item{snowenv}{one of \code{Alpine}, \code{Maritime}, \code{Prairie}, \code{Taiga} or \code{Tundra}}
}
\value{
a single numeric value indicating the melt factor
}
\description{
The function \code{getmeltf} returns pre-calculated snow melt parameters,
derived using \code{\link[=fitsnowdepth]{fitsnowdepth()}} for different snow environments.
}
