% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgt}
\alias{hgt}
\title{A 25 m resolution dataset of canopy height.}
\format{
A PackedSpatRaster object with 50 rows and 50 columns
}
\source{
\url{https://land.copernicus.eu/}
}
\usage{
hgt
}
\description{
A dataset of canopy heights (m) for Sodankylä in Finland (xmin = 2963700, xmax = 2964950,
ymin = 10221300, ymax = 10222550, CRS: 3395).
}
\keyword{datasets}
