% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{modelsnowdepth}
\alias{modelsnowdepth}
\title{Runs spatial snow model}
\usage{
modelsnowdepth(
  weather,
  precd,
  snowdepth,
  dtm,
  slr = NA,
  apr = NA,
  pai,
  hgt,
  STparams,
  meltfact = NA,
  plai = 0.3,
  x = 1,
  lat = NA,
  long = NA,
  snowenv = "Taiga",
  tpi_radius = 200,
  tfact = 10,
  snowem = 0.99,
  zmin = 0.002,
  umin = 0.5,
  astc = 1.5,
  spatialmelt = FALSE,
  Sh = 6.3,
  zu = 2,
  xyf = NA,
  merid = 0,
  dst = 0,
  initdepth = 0,
  out = "hourly"
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{precd}{a vector of daily precipitation (mm).}

\item{snowdepth}{a numeric vector of length equal to the number of timesteps (e.g. nrow(weather)) representing average snowdepth, in cm, across the scene}

\item{dtm}{a raster of elevations (m). the x and y dimensions of the raster must also be in metres}

\item{slr}{an optional raster object of slope values (Radians). Calculated from dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional raster object of aspect values (Radians). Calculated from dtm if not supplied, but outer cells will be NA.}

\item{pai}{a single numeric value, raster or array of plant area index values}

\item{hgt}{a raster of vegetation heights}

\item{STparams}{snow temperature model coefficients as derived by \code{\link[=fitsnowtemp]{fitsnowtemp()}}}

\item{meltfact}{snow melt coefficient as returned by \code{fitsnowtemp} or \code{getmeltf}. Derived using \code{getmeltf} is not supplied.}

\item{plai}{a single numeric value, raster or array of the proprotion of plant area index values that are leaves}

\item{x}{optional single numeric value, raster, matrix or array of leaf distribution coefficients}

\item{lat}{latitude of location (decimal degrees). Derived from \code{dtm} is not supplied, so coordinate reference of system of \code{dtm} must be defined.}

\item{long}{longitude of location (decimal degrees). Derived from \code{dtm} is not supplied, so coordinate reference of system of \code{dtm} must be defined.}

\item{snowenv}{one of \code{Alpine}, \code{Maritime}, \code{Prairie}, \code{Taiga} or \code{Tundra} (see details)}

\item{tpi_radius}{radius for applying topographic positioning index when distributing snow as returned by \code{\link[=tpiradius]{tpiradius()}}}

\item{tfact}{single numeric value specifying sensitivity of snow distribution to topographic positioning index}

\item{snowem}{optionally, numeric value of snow emissivity}

\item{zmin}{optionally, numeric value of roughness length for momentum transfer of snow surface without vegetation protruding  (m)}

\item{umin}{optionally, numeric value indicating minimum wind speed used in conductivity calculations (m/s)}

\item{astc}{optionally, numeric value indicating the temperature at which precipitation falls as snow (deg C)}

\item{spatialmelt}{optional logical indicating whether or not snow melt varies spatially (e.g. greater on sun-facing slopes). Model takes longer to run if \code{TRUE}}

\item{Sh}{optionally, branch snow load coefficient (kg/m^2). 6.6 for pine and 5.9 kg for spruce.}

\item{zu}{height above ground of wind speed measurement in \code{weather} (m)}

\item{xyf}{number of grid cells over which to smooth vertical wind height profile}

\item{merid}{longitude of local timezone meridian (0 for UTC)}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for British Summer Time if \code{merid} = 0).}

\item{initdepth}{single numeric value or matrix of initial snow depths at start of model run}

\item{out}{optinal variable indicating whether to return hourly or daily snow depths (default is hourly)}
}
\value{
a list of the following:

\code{gsnowdepth} array of predicted ground snow depth (cm) across the scene

\code{canswe} array of predicted canopy snow water equivalent (mm / m^2) across the scene

\code{snowtempG} array of predicted ground snow surface temperature (deg C) across the scene
}
\description{
\code{modelsnowdepth} runs the spatial snow model to derive an
array of snow depths and temperatures for each time increment in \code{weather}
and over each grid cell of \code{dtm}
}
\details{
#' @details The format and and units of \code{weather} must follow that in the example
dataset \code{climdata}. The paramater \code{snowenv} is used to compute snow density
following Sturm et al (2010) J Hydrometeorology 11: 1380-1393. The leaf distribution angle
coefficient is the ratio of vertical to horizontal projections of leaf foliage (~1 for decidious woodland).
}
\examples{
require(NicheMapR)
# Derive estimates of snow melt temperature parameters using NicheMapR
nmrout<-runNMRSnow(climdata,precd,66.215,29.293,ALTT = 360)
STparams<-fitsnowtemp(climdata,precd,nmrout$SNOWDEP,nmrout$SNOWTEMP)
meltfact<-fitsnowdepth(climdata,nmrout$SNOWDEP,precd,STparams=STparams)$meltfact
# Run snow model with defaults (takes 20 seconds)
mout1<-modelsnowdepth(climdata,precd,dtm,pai,hgt,STparams,meltfact)
# Run snow model with defaults, but allowing spatially variable snow melt (takes ~100 seconds)
mout2<-modelsnowdepth(climdata,precd,dtm,pai,hgt,STparams,meltfact,spatialmelt = TRUE)
# Compare results
msnowdepth1<-apply(mout1$gsnowdepth,c(1,2),mean)
msnowdepth2<-apply(mout2$gsnowdepth,c(1,2),mean)
snowdif<-msnowdepth2-msnowdepth1
par(mfrow=c(2,2))
plot(raster(msnowdepth1))
plot(raster(msnowdepth2))
plot(raster(snowdif))
}
