% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pai}
\alias{pai}
\title{A 25 m resolution dataset of plant area index values.}
\format{
A PackedSpatRaster object with 50 rows and 50 columns
}
\source{
\url{https://land.copernicus.eu/}
}
\usage{
pai
}
\description{
A dataset of plant area index values for Sodankylä in Finland (xmin = 2963700, xmax = 2964950,
ymin = 10221300, ymax = 10222550, CRS: 3395).
}
\keyword{datasets}
