% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precd}
\alias{precd}
\title{Daily precipitation}
\format{
A vector of daily precipitation (mm/day, snow water equivalent when snow)
}
\source{
\url{https://en.ilmatieteenlaitos.fi/}
}
\usage{
precd
}
\description{
A vector of daily precipitation (2nd Oct 2017 to 1st Apr 2019) for Sodankylä in Finland (67.367N, 26.629E)
}
\keyword{datasets}
