% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowfitfunctions.R
\name{pSnow}
\alias{pSnow}
\title{Estimates snow depth and other snow variables}
\usage{
pSnow(
  weather,
  precd,
  snowenv = "Taiga",
  STparams,
  meltfact = 0.11,
  snowem = 0.99,
  zm = 0.002,
  umin = 0.5,
  astc = 1.5,
  initdepth = 0
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{precd}{a vector of daily precipitation (mm).}

\item{snowenv}{one of \code{Alpine}, \code{Maritime}, \code{Prairie}, \code{Taiga} or \code{Tundra} (see details)}

\item{STparams}{an object of class \code{SnowTparams} as returned by \code{\link[=fitsnowtemp]{fitsnowtemp()}}}

\item{meltfact}{a snow melt coefficient, as returned by \code{\link[=fitsnowdepth]{fitsnowdepth()}} or \code{\link[=getmeltf]{getmeltf()}}}

\item{snowem}{optionally, snow emissivity}

\item{zm}{optionally, roughness length for momentum transfer of snow surface (m)}

\item{umin}{optionally, numeric value indicating minimum wind speed used in conductivity calculations (m/s)}

\item{astc}{optionally, numeric value indicating the temperature at which precipitation falls as snow (deg C)}

\item{initdepth}{single numeric value giving the initial snow depth (cm)}
}
\value{
a list of the following:

\code{snowdepth} snow depth (cm)

\code{snowdens} snow density (g/cm^3)

\code{snowtemp} snow temperature (deg C)

\code{snowref} snow reflectance (range 0 to 1)

\code{snowfall} snow fall (cm) at each time increment

\code{snowmelt} snow melt (cm) at each time increment
}
\description{
The function \code{pSnow} estimates snow depth, albedo, density and temperature
}
\details{
The format and and units of \code{weather} must follow that in the example
dataset \code{climdata}. The paramater \code{snowenv} is used to compute snow density
following Sturm et al (2010) J Hydrometeorology 11: 1380-1393.
}
