% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowfitfunctions.R
\name{runNMRSnow}
\alias{runNMRSnow}
\title{Wrapper function for running NicheMapR snow model}
\usage{
runNMRSnow(
  weather,
  precd,
  lat,
  long,
  snowenv = "Taiga",
  zm = 0.002,
  snowem = 0.99,
  ALTT = 0
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{precd}{a vector of daily precipitation (mm).}

\item{lat}{latitude of location for which snow depth and temperature are required (decimal degrees)}

\item{long}{longitude of location for which snow depth and temperature are required (decimal degrees)}

\item{snowenv}{one of \code{Alpine}, \code{Maritime}, \code{Prairie}, \code{Taiga} or \code{Tundra} (see details)}

\item{zm}{optionally, numeric value of roughness length for momentum transfer of snow surface (m)}

\item{snowem}{optionally, snow emissivity (~0.99) speed used in conductivity calculations (m/s)}

\item{ALTT}{elevation of location (m)}
}
\value{
a list of the following:

\code{SNOWDEP} vector of predicted snow depth (cm)

\code{SNOWTEMP} vector of predicted snow surface temperature (deg C)
}
\description{
The function \code{runNMRSnow} runs the multi-layer snow model in \code{NicheMapR}
(Kearney and POrter 2017, Ecography 40: 664-674) and can be used
to derive hourly snow depths and temperatures for fitting the simpler model emulator
in \code{microsnow}. Requires \code{NicheMapR} to be installed (https://github.com/mrke/NicheMapR)
}
\details{
The format and and units of \code{weather} must follow that in the example
dataset \code{climdata}. The paramater \code{snowenv} is used to compute snow density
following Sturm et al (2010) J Hydrometeorology 11: 1380-1393.
}
