% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SDparams}
\alias{SDparams}
\title{an object of class \code{SnowDparams}}
\format{
a list of the following outputs:
\describe{
\item{psnowdepth}{A vector of predicted snow depths}
\item{RMS}{RMS error of fitted model}
\item{meltfac}{A snow melt coefficient}
}
}
\usage{
SDparams
}
\description{
A list of model parameters for modelling snow depth (here derived using \code{\link[=fitsnowdepth]{fitsnowdepth()}} on inbuilt datasets)
}
\keyword{datasets}
