% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{snowbelow_hr}
\alias{snowbelow_hr}
\title{Below ground microclimate model}
\usage{
snowbelow_hr(
  microsnow,
  reqhgt,
  pai_a = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA
)
}
\arguments{
\item{microsnow}{an object of class microsnowin as returned by \code{\link[=snowmodelin]{snowmodelin()}}}

\item{reqhgt}{specified height (m) below ground for which microclimate
outputs are required. Must be negative.}

\item{pai_a}{total plant area index above \code{reqhgt}. Estimated if not supplied (see details).}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from
dtm if not supplied.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
dtm if not supplied.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied.}
}
\value{
an array of soil temperatures (deg C)
}
\description{
Used by e.g. \code{\link[=runmicro_hr]{runmicro_hr()}} to model microclimate below ground
}
