% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snowdepth}
\alias{snowdepth}
\title{A data frame of hourly snow depths}
\format{
a data frame with the following elements:
\describe{
\item{obs_time}{POSIXlt object of dates and times in UTC}
\item{snowdepth}{snow dpeth (cm)}
}
}
\source{
\url{https://en.ilmatieteenlaitos.fi/}
}
\usage{
snowdepth
}
\description{
A data frame of hourly snow depths (2nd Oct 2017 to 1st Apr 2019) for Sodankylä in Finland (67.367N, 26.629E)
}
\keyword{datasets}
