% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowfitfunctions.R
\name{SnowTemp}
\alias{SnowTemp}
\title{Estimates snow temperature}
\usage{
SnowTemp(Rabs, gHa, tc, STparams, snowem = 0.99)
}
\arguments{
\item{Rabs}{radiation absorbed by snow surface (W/m^2)}

\item{gHa}{molar conductivity for heat (mol/m^2/s)}

\item{tc}{air temperature (deg G)}

\item{STparams}{an object of class \code{SnowTparams} as returned by \code{\link[=fitsnowtemp]{fitsnowtemp()}}}

\item{snowem}{optionally, numeric value(s) of snow emissivity (~0.95)}
}
\value{
snow temperature (deg C)
}
\description{
The function \code{SnowTemp} estimates snow temperatures
}
