% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STparams}
\alias{STparams}
\title{an object of class \code{SnowTparams}}
\format{
a list of the following outputs:
\describe{
\item{m1}{A list of model coefficients when snow temperature below freezing}
\item{m2}{A list of model coefficients when snow temperature above freezing}
}
}
\usage{
STparams
}
\description{
A list of model parameters for modelling snow temperatures (here derived using \code{\link[=fitsnowtemp]{fitsnowtemp()}} on inbuilt datasets)
}
\keyword{datasets}
