% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{tpiradius}
\alias{tpiradius}
\title{Calculates appropriate radius over which to apply topographic position index
NB function needs tuning}
\usage{
tpiradius(windspeed, hgt, PAI)
}
\description{
Calculates appropriate radius over which to apply topographic position index
NB function needs tuning
}
