% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdistfunctions.R
\name{.snowrad}
\alias{.snowrad}
\title{Calculates radiation absorbed by snow surface}
\usage{
.snowrad(
  weather,
  snowdepth,
  dtm,
  hor,
  lat,
  long,
  epai,
  x,
  snowalb,
  snowem,
  slr = NA,
  apr = NA,
  clump = 0,
  clumpd = "daily"
)
}
\description{
Calculates radiation absorbed by snow surface
}
